Remove system tests covered by unit tests
